chore: working go builds
