+++ packages/go/stbernard/command/builder/builder.go
@@ -1,0 +1,60 @@
+package builder
+
+import (
+	"flag"
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"github.com/specterops/bloodhound/packages/go/stbernard/workspace"
+)
+
+const (
+	Name  = "build"
+	Usage = "Build commands in current workspace"
+)
+
+type Config struct {
+	Environment []string
+}
+
+type command struct {
+	config Config
+}
+
+func (s command) Usage() string {
+	return Usage
+}
+
+func (s command) Name() string {
+	return Name
+}
+
+func (s command) Run() error {
+	if cwd, err := workspace.FindRoot(); err != nil {
+		return fmt.Errorf("could not find workspace root: %w", err)
+	} else if modPaths, err := workspace.ParseModulesAbsPaths(cwd); err != nil {
+		return fmt.Errorf("could not parse module absolute paths: %w", err)
+	} else if err := workspace.BuildMainPackages(cwd, modPaths); err != nil {
+		return fmt.Errorf("could not build main packages: %w", err)
+	} else {
+		return nil
+	}
+}
+
+func Create(config Config) (command, error) {
+	cmd := flag.NewFlagSet(Name, flag.ExitOnError)
+
+	cmd.Usage = func() {
+		w := flag.CommandLine.Output()
+		fmt.Fprintf(w, "%s\n\nUsage: %s %s [OPTIONS]\n\nOptions:\n", Usage, filepath.Base(os.Args[0]), Name)
+		cmd.PrintDefaults()
+	}
+
+	if err := cmd.Parse(os.Args[2:]); err != nil {
+		cmd.Usage()
+		return command{}, fmt.Errorf("failed to parse build command: %w", err)
+	} else {
+		return command{config: config}, nil
+	}
+}
